🗂️ Reorganize performance testing into centralized structure
📁 New Structure:
• performance/scripts/k6/ - K6 load testing scripts
• performance/scripts/shell/ - Shell orchestration scripts
• performance/monitoring/ - Real-time monitoring tools
• performance/results/ - Test results (current + archive)
• performance/docs/ - Performance testing documentation

✨ Features:
• Comprehensive performance testing README
• Updated script paths and references
• Centralized results management
• Better organization for scalability

🚀 Usage: ./performance/scripts/shell/find-breaking-point.sh
