+++ performance/results/archive/results/monitoring_20250903_090308.log
@@ -1,0 +1,39 @@
+🔍 Starting Performance Monitoring for 180 seconds...
+📊 Results will be saved to: docs/performance/results/metrics_20250903_090308.txt
+📈 Collecting metrics (1/6)...
+Wed Sep  3 09:03:08 -03 2025: Collecting metrics...
+jq: error (at <stdin>:0): Cannot iterate over null (null)
+⏳ Waiting 30 seconds...
+📈 Collecting metrics (2/6)...
+Wed Sep  3 09:03:39 -03 2025: Collecting metrics...
+jq: error (at <stdin>:0): Cannot iterate over null (null)
+⏳ Waiting 30 seconds...
+📈 Collecting metrics (3/6)...
+Wed Sep  3 09:04:09 -03 2025: Collecting metrics...
+jq: error (at <stdin>:0): Cannot iterate over null (null)
+⏳ Waiting 30 seconds...
+📈 Collecting metrics (4/6)...
+Wed Sep  3 09:04:40 -03 2025: Collecting metrics...
+jq: error (at <stdin>:0): Cannot iterate over null (null)
+⏳ Waiting 30 seconds...
+📈 Collecting metrics (5/6)...
+Wed Sep  3 09:05:10 -03 2025: Collecting metrics...
+jq: error (at <stdin>:0): Cannot iterate over null (null)
+⏳ Waiting 30 seconds...
+📈 Collecting metrics (6/6)...
+Wed Sep  3 09:05:40 -03 2025: Collecting metrics...
+jq: error (at <stdin>:0): Cannot iterate over null (null)
+✅ Monitoring completed!
+📄 Results saved to: docs/performance/results/metrics_20250903_090308.txt
+
+📊 SUMMARY
+==========
+Duration: 180 seconds
+Samples: 6
+Output: docs/performance/results/metrics_20250903_090308.txt
+
+🔍 To view results:
+cat docs/performance/results/metrics_20250903_090308.txt
+
+📈 To analyze with Prometheus directly:
+curl 'http://localhost:9090/api/v1/query?query=rate(wallet_operations_deposits_total[5m])'
