+++ performance/results/current/stress-test-20250904_075239/baseline-cpu.json
@@ -1,0 +1,1 @@
+{"status":"error","errorType":"bad_data","error":"invalid parameter \"query\": 1:21: parse error: unexpected \"=\""}
