+++ performance/results/current/stress-test-20250904_075239/baseline-memory.json
@@ -1,0 +1,1 @@
+{"status":"error","errorType":"bad_data","error":"invalid parameter \"query\": 1:25: parse error: unexpected \"=\""}{"status":"error","errorType":"bad_data","error":"invalid parameter \"query\": 1:26: parse error: unexpected \"=\""}
