+++ results/breaking-point-20250904_080648/test-output.log
@@ -1,0 +1,132 @@
+
+         /\      Grafana   /‾‾/  
+    /\  /  \     |\  __   /  /   
+   /  \/    \    | |/ /  /   ‾‾\ 
+  /          \   |   (  |  (‾)  |
+ / __________ \  |_|\_\  \_____/ 
+
+     execution: local
+        script: scripts/continuous-wallet-stress.js
+        output: json (results/breaking-point-20250904_080648/results.json)
+
+     scenarios: (100.00%) 1 scenario, 500 max VUs, 23m30s max duration (incl. graceful stop):
+              * default: Up to 500 looping VUs for 23m0s over 12 stages (gracefulRampDown: 30s, gracefulStop: 30s)
+
+time="2025-09-04T08:06:48-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (15ms)" source=console
+
+running (00m01.0s), 001/500 VUs, 0 complete and 0 interrupted iterations
+default   [   0% ] 001/500 VUs  00m01.0s/23m00.0s
+
+running (00m02.0s), 001/500 VUs, 0 complete and 0 interrupted iterations
+default   [   0% ] 001/500 VUs  00m02.0s/23m00.0s
+time="2025-09-04T08:06:50-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (13ms)" source=console
+
+running (00m03.0s), 001/500 VUs, 1 complete and 0 interrupted iterations
+default   [   0% ] 001/500 VUs  00m03.0s/23m00.0s
+
+running (00m04.0s), 001/500 VUs, 1 complete and 0 interrupted iterations
+default   [   0% ] 001/500 VUs  00m04.0s/23m00.0s
+time="2025-09-04T08:06:52-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (15ms)" source=console
+
+running (00m05.0s), 001/500 VUs, 2 complete and 0 interrupted iterations
+default   [   0% ] 001/500 VUs  00m05.0s/23m00.0s
+
+running (00m06.0s), 001/500 VUs, 2 complete and 0 interrupted iterations
+default   [   0% ] 001/500 VUs  00m06.0s/23m00.0s
+time="2025-09-04T08:06:54-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (12ms)" source=console
+
+running (00m07.0s), 001/500 VUs, 3 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m07.0s/23m00.0s
+
+running (00m08.0s), 001/500 VUs, 3 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m08.0s/23m00.0s
+time="2025-09-04T08:06:56-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (16ms)" source=console
+
+running (00m09.0s), 001/500 VUs, 4 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m09.0s/23m00.0s
+
+running (00m10.0s), 001/500 VUs, 4 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m10.0s/23m00.0s
+time="2025-09-04T08:06:58-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (15ms)" source=console
+
+running (00m11.0s), 001/500 VUs, 5 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m11.0s/23m00.0s
+
+running (00m12.0s), 001/500 VUs, 5 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m12.0s/23m00.0s
+time="2025-09-04T08:07:00-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (19ms)" source=console
+
+running (00m13.0s), 001/500 VUs, 6 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m13.0s/23m00.0s
+
+running (00m14.0s), 001/500 VUs, 6 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m14.0s/23m00.0s
+time="2025-09-04T08:07:02-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (14ms)" source=console
+
+running (00m15.0s), 001/500 VUs, 7 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m15.0s/23m00.0s
+
+running (00m16.0s), 001/500 VUs, 7 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m16.0s/23m00.0s
+time="2025-09-04T08:07:04-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (14ms)" source=console
+
+running (00m17.0s), 001/500 VUs, 8 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m17.0s/23m00.0s
+
+running (00m18.0s), 001/500 VUs, 8 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m18.0s/23m00.0s
+time="2025-09-04T08:07:06-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (19ms)" source=console
+
+running (00m19.0s), 001/500 VUs, 9 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m19.0s/23m00.0s
+
+running (00m20.0s), 001/500 VUs, 9 complete and 0 interrupted iterations
+default   [   1% ] 001/500 VUs  00m20.0s/23m00.0s
+time="2025-09-04T08:07:08-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (16ms)" source=console
+
+running (00m21.0s), 001/500 VUs, 10 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m21.0s/23m00.0s
+
+running (00m22.0s), 001/500 VUs, 10 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m22.0s/23m00.0s
+time="2025-09-04T08:07:10-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (15ms)" source=console
+
+running (00m23.0s), 001/500 VUs, 11 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m23.0s/23m00.0s
+
+running (00m24.0s), 001/500 VUs, 11 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m24.0s/23m00.0s
+time="2025-09-04T08:07:12-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (28ms)" source=console
+
+running (00m25.0s), 001/500 VUs, 12 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m25.0s/23m00.0s
+
+running (00m26.0s), 001/500 VUs, 12 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m26.0s/23m00.0s
+time="2025-09-04T08:07:14-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (16ms)" source=console
+
+running (00m27.0s), 001/500 VUs, 13 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m27.0s/23m00.0s
+
+running (00m28.0s), 001/500 VUs, 13 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m28.0s/23m00.0s
+time="2025-09-04T08:07:16-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (13ms)" source=console
+
+running (00m29.0s), 001/500 VUs, 14 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m29.0s/23m00.0s
+
+running (00m30.0s), 001/500 VUs, 14 complete and 0 interrupted iterations
+default   [   2% ] 001/500 VUs  00m30.0s/23m00.0s
+time="2025-09-04T08:07:18-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (10ms)" source=console
+time="2025-09-04T08:07:18-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (21ms)" source=console
+
+running (00m31.0s), 002/500 VUs, 15 complete and 0 interrupted iterations
+default   [   2% ] 002/500 VUs  00m31.0s/23m00.0s
+
+running (00m32.0s), 002/500 VUs, 15 complete and 0 interrupted iterations
+default   [   2% ] 002/500 VUs  00m32.0s/23m00.0s
+time="2025-09-04T08:07:20-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (5ms)" source=console
+time="2025-09-04T08:07:20-03:00" level=info msg="❌ Failed to create wallet: HTTP 201 (13ms)" source=console
+
+running (00m33.0s), 002/500 VUs, 17 complete and 0 interrupted iterations
+default   [   2% ] 002/500 VUs  00m33.0s/23m00.0s
